商户认证和管理

+++ 精灵之泉/src/main/java/com/sxbwstxpay/model/OrderVippay.java
@@ -80,7 +80,7 @@ public class OrderVippay {
 
         private String pay_info;
         private String payment_name;
-        private int pay_money;
+        private double pay_money;
         private String class_name;
         private ConfigBean config;
 
@@ -100,11 +100,11 @@ public class OrderVippay {
             this.payment_name = payment_name;
         }
 
-        public int getPay_money() {
+        public double getPay_money() {
             return pay_money;
         }
 
-        public void setPay_money(int pay_money) {
+        public void setPay_money(double pay_money) {
             this.pay_money = pay_money;
         }
 
@@ -157,8 +157,8 @@ public class OrderVippay {
             private String packagevalue;
             private String subject;
             private Object body;
-            private int total_fee;
-            private int total_fee_format;
+            private double total_fee;
+            private double total_fee_format;
             private Object out_trade_no;
             private String notify_url;
             private String key;
@@ -252,19 +252,19 @@ public class OrderVippay {
                 this.body = body;
             }
 
-            public int getTotal_fee() {
+            public double getTotal_fee() {
                 return total_fee;
             }
 
-            public void setTotal_fee(int total_fee) {
+            public void setTotal_fee(double total_fee) {
                 this.total_fee = total_fee;
             }
 
-            public int getTotal_fee_format() {
+            public double getTotal_fee_format() {
                 return total_fee_format;
             }
 
-            public void setTotal_fee_format(int total_fee_format) {
+            public void setTotal_fee_format(double total_fee_format) {
                 this.total_fee_format = total_fee_format;
             }
